[win32] filesystem: move implementation to platform folder
