+++ cmake/treedata/windows/subdirs.txt
@@ -1,17 +1,17 @@
-xbmc/platform/win32          platform/win32
-xbmc/platform/win32/input    platworm/win32/input
-xbmc/input/touch             input/touch
-xbmc/input/touch/generic     input/touch/generic
-xbmc/network/windows         network/windows
-xbmc/network/mdns            network/mdns
-xbmc/peripherals/bus/win32   peripherals/bus/win32
-xbmc/powermanagement/windows powermanagement/windows
-xbmc/storage/windows         storage/windows
-xbmc/filesystem/win32        filesystem/win32
-xbmc/utils/win32             utils/win32
-xbmc/rendering/dx            rendering/dx
-xbmc/threads/platform/win    threads/win
-xbmc/windowing/windows       windowing/windows
+xbmc/platform/win32             platform/win32
+xbmc/platform/win32/input       platworm/win32/input
+xbmc/platform/win32/filesystem  platform/win32/filesystem
+xbmc/input/touch                input/touch
+xbmc/input/touch/generic        input/touch/generic
+xbmc/network/windows            network/windows
+xbmc/network/mdns               network/mdns
+xbmc/peripherals/bus/win32      peripherals/bus/win32
+xbmc/powermanagement/windows    powermanagement/windows
+xbmc/storage/windows            storage/windows
+xbmc/utils/win32                utils/win32
+xbmc/rendering/dx               rendering/dx
+xbmc/threads/platform/win       threads/win
+xbmc/windowing/windows          windowing/windows
 xbmc/cores/RetroPlayer/process/windows cores/RetroPlayer/process/windows
 xbmc/cores/RetroPlayer/rendering/VideoShaders/windows cores/RetroPlayer/rendering/VideoShaders/windows
 xbmc/cores/VideoPlayer/Process/windows cores/VideoPlayer/Process/windows
