+++ xbmc/platform/win32/filesystem/Win32Directory.h
