+++ xbmc/platform/win32/filesystem/Win32File.cpp
