+++ xbmc/platform/win32/filesystem/Win32SMBDirectory.cpp
