+++ xbmc/platform/win32/filesystem/Win32SMBFile.h
